// src/infra/whatsapp-client/entity/whatsapp-client-users-menu.enum.ts
export enum WhatsappClientUsersMenu {
  INITIAL = 'initial',
  LATEST_NEWS = 'latest_news',
  MATCH_SCHEDULE = 'match_schedule',
  TROPHY_HISTORY = 'trophy_history',
  MATCH_ALERTS = 'match_alerts',
  LIVE_GAME_STATUS = 'live_game_status',
  FURIA_AI_CHAT = 'furia_ai_chat',
  OFFICIAL_STORE = 'official_store',
}